split out common Actor tasks
