+++ app/actors/hyrax/actors/image_actor.rb
@@ -1,10 +1,7 @@
 # frozen_string_literal: true
-
-# Generated via
-#  `rails generate hyrax:work Image`
 module Hyrax
   module Actors
-    class ImageActor < Hyrax::Actors::BaseActor
+    class ImageActor < ::Spot::BaseActor
     end
   end
 end
