+++ app/forms/hyrax/image_form.rb
@@ -1,10 +1,11 @@
 # frozen_string_literal: true
-
-# Generated via
-#  `rails generate hyrax:work Image`
 module Hyrax
   # Generated form for Image
   class ImageForm < Hyrax::Forms::WorkForm
+    include ::IdentifierFormFields
+    include ::LanguageTaggedFormFields
+    include ::NestedFormFields
+
     self.model_class = ::Image
     self.terms += [:resource_type]
   end
