+++ spec/support/shared_examples/spot_presenter.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
-RSpec.shared_examples 'a spot presenter' do
+RSpec.shared_examples 'a Spot presenter' do
   subject(:presenter) { described_class.new(solr_doc, ability) }
 
   let(:factory) { described_class.name.split('::').last.gsub(/Presenter/, '').downcase.to_sym }
